Enabled Multiplayer Feature
This enables the multiplayer feature. When the game is started, the user
is asked if he/she wants to play single player or multiplayer.

For multiplayer: Player One uses arrow keys, Player Two uses WASD.

The issue is that when one player eats food and then the other player
eats food, the food disappears. (It is still there but it turns white)

Also, changing the colors of the two players may be a good idea.