Restart, variable names, Sound effect
In the terminal, I added something asking whether you want to play again or stop. The only error is that the previously played snake continues alongside the new snake. I will try to fix that soon.  I also fixed the variable names to more specific ones(example: variable "c" is changed to colorObject and "C" is changed to ColorArray.
I also added a sound effect for whenever the snake eats the apple.